+++ changelog.md
@@ -2,7 +2,9 @@
 
 ## UNRELEASED
 
-_Nothing yet..._
+**Added:**
+- react-native-bpk-component-card: 0.0.3 => 1.0.0
+  - New component BpkCard, see http://backpack.prod.aws.skyscnr.com/components/native/cards/
 
 ## 2017-09-26 - New native Card component